+++ packages/backend/src/services/UserService.ts
@@ -142,6 +142,9 @@ export class UserService {
     }
 
     async delete(user: SessionUser, userUuidToDelete: string): Promise<void> {
+        if (user.userUuid === userUuidToDelete) {
+            throw new ForbiddenError('User can not delete themself');
+        }
         if (user.organizationUuid === undefined) {
             throw new NotExistsError('Organization not found');
         }
