+++ packages/common/src/index.ts
@@ -46,6 +46,19 @@ import { ShareUrl } from './types/share';
 import { SlackSettings } from './types/slackSettings';
 
 import { FieldValueSearchResult } from './types/fieldMatch';
+import {
+    OnboardingStatus,
+    Organisation,
+    OrganizationProject,
+} from './types/organization';
+import {
+    CreateWarehouseCredentials,
+    DbtProjectConfig,
+    DbtProjectType,
+    Project,
+    ProjectType,
+    WarehouseCredentials,
+} from './types/projects';
 import { SchedulerAndTargets } from './types/scheduler';
 import { SlackChannel } from './types/slack';
 import { Space } from './types/space';
@@ -53,7 +66,6 @@ import { TableBase } from './types/table';
 import { LightdashUser } from './types/user';
 import { formatItemValue } from './utils/formatting';
 import { getItemId, getItemLabelWithoutTableName } from './utils/item';
-import { WeekDay } from './utils/timeFrames';
 
 export * from './authorization/index';
 export * from './authorization/types';
@@ -84,6 +96,7 @@ export * from './types/organizationMemberProfile';
 export * from './types/personalAccessToken';
 export * from './types/pinning';
 export * from './types/projectMemberProfile';
+export * from './types/projects';
 export * from './types/results';
 export * from './types/savedCharts';
 export * from './types/scheduler';
@@ -154,21 +167,6 @@ export function hexToRGB(hex: string, alpha: number | undefined): string {
         return `rgba(${r}, ${g}, ${b}, ${alpha})`;
     }
     return `rgb(${r}, ${g}, ${b})`;
-}
-
-export enum ProjectType {
-    DEFAULT = 'DEFAULT',
-    PREVIEW = 'PREVIEW',
-}
-
-export enum DbtProjectType {
-    DBT = 'dbt',
-    DBT_CLOUD_IDE = 'dbt_cloud_ide',
-    GITHUB = 'github',
-    GITLAB = 'gitlab',
-    BITBUCKET = 'bitbucket',
-    AZURE_DEVOPS = 'azure_devops',
-    NONE = 'none',
 }
 
 // Seeds
@@ -425,28 +423,9 @@ export type CreateInviteLink = Pick<InviteLink, 'expiresAt' | 'email'> & {
     role?: OrganizationMemberRole;
 };
 
-export type OnbordingRecord = {
-    ranQueryAt: Date | null;
-    shownSuccessAt: Date | null;
-};
-
-export type OnboardingStatus = {
-    ranQuery: boolean;
-};
-
 export type ProjectSavedChartStatus = boolean;
 
 export type ApiFlashResults = Record<string, string[]>;
-
-export type Organisation = {
-    organizationUuid: string;
-    name: string;
-    allowedEmailDomains: string[];
-    chartColors?: string[];
-    needsProject?: boolean;
-};
-
-export type UpdateOrganisation = Partial<Organisation>;
 
 type ApiResults =
     | ApiQueryResults
@@ -593,163 +572,10 @@ export enum DBFieldTypes {
     METRIC = 'metric',
 }
 
-export enum WarehouseTypes {
-    BIGQUERY = 'bigquery',
-    POSTGRES = 'postgres',
-    REDSHIFT = 'redshift',
-    SNOWFLAKE = 'snowflake',
-    DATABRICKS = 'databricks',
-    TRINO = 'trino',
-}
-
-export type CreateBigqueryCredentials = {
-    type: WarehouseTypes.BIGQUERY;
-    project: string;
-    dataset: string;
-    threads?: number;
-    timeoutSeconds: number | undefined;
-    priority: 'interactive' | 'batch' | undefined;
-    keyfileContents: Record<string, string>;
-    retries: number | undefined;
-    location: string | undefined;
-    maximumBytesBilled: number | undefined;
-    startOfWeek?: WeekDay | null;
-};
-
-export const sensitiveCredentialsFieldNames = [
-    'user',
-    'password',
-    'keyfileContents',
-    'personalAccessToken',
-    'privateKey',
-    'privateKeyPass',
-] as const;
-
 export const sensitiveDbtCredentialsFieldNames = [
     'personal_access_token',
     'api_key',
 ] as const;
-
-export type SensitiveCredentialsFieldNames =
-    typeof sensitiveCredentialsFieldNames[number];
-
-export type BigqueryCredentials = Omit<
-    CreateBigqueryCredentials,
-    SensitiveCredentialsFieldNames
->;
-
-export type CreateDatabricksCredentials = {
-    type: WarehouseTypes.DATABRICKS;
-    catalog?: string;
-    // this supposed to be a `schema` but changing it will break for existing customers
-    database: string;
-    serverHostName: string;
-    httpPath: string;
-    personalAccessToken: string;
-    startOfWeek?: WeekDay | null;
-};
-
-export type DatabricksCredentials = Omit<
-    CreateDatabricksCredentials,
-    SensitiveCredentialsFieldNames
->;
-
-export type CreatePostgresCredentials = {
-    type: WarehouseTypes.POSTGRES;
-    host: string;
-    user: string;
-    password: string;
-    port: number;
-    dbname: string;
-    schema: string;
-    threads?: number;
-    keepalivesIdle?: number;
-    searchPath?: string;
-    role?: string;
-    sslmode?: string;
-    startOfWeek?: WeekDay | null;
-};
-
-export type PostgresCredentials = Omit<
-    CreatePostgresCredentials,
-    SensitiveCredentialsFieldNames
->;
-
-export type CreateTrinoCredentials = {
-    type: WarehouseTypes.TRINO;
-    host: string;
-    user: string;
-    password: string;
-    port: number;
-    dbname: string;
-    schema: string;
-    http_scheme: string;
-    startOfWeek?: WeekDay | null;
-};
-
-export type TrinoCredentials = Omit<
-    CreateTrinoCredentials,
-    SensitiveCredentialsFieldNames
->;
-
-export type CreateRedshiftCredentials = {
-    type: WarehouseTypes.REDSHIFT;
-    host: string;
-    user: string;
-    password: string;
-    port: number;
-    dbname: string;
-    schema: string;
-    threads?: number;
-    keepalivesIdle?: number;
-    sslmode?: string;
-    ra3Node?: boolean;
-    startOfWeek?: WeekDay | null;
-};
-
-export type RedshiftCredentials = Omit<
-    CreateRedshiftCredentials,
-    SensitiveCredentialsFieldNames
->;
-
-export type CreateSnowflakeCredentials = {
-    type: WarehouseTypes.SNOWFLAKE;
-    account: string;
-    user: string;
-    password?: string;
-    privateKey?: string;
-    privateKeyPass?: string;
-    role?: string;
-    database: string;
-    warehouse: string;
-    schema: string;
-    threads?: number;
-    clientSessionKeepAlive?: boolean;
-    queryTag?: string;
-    accessUrl?: string;
-    startOfWeek?: WeekDay | null;
-};
-
-export type SnowflakeCredentials = Omit<
-    CreateSnowflakeCredentials,
-    SensitiveCredentialsFieldNames
->;
-
-export type CreateWarehouseCredentials =
-    | CreateRedshiftCredentials
-    | CreateBigqueryCredentials
-    | CreatePostgresCredentials
-    | CreateSnowflakeCredentials
-    | CreateDatabricksCredentials
-    | CreateTrinoCredentials;
-
-export type WarehouseCredentials =
-    | SnowflakeCredentials
-    | RedshiftCredentials
-    | PostgresCredentials
-    | BigqueryCredentials
-    | DatabricksCredentials
-    | TrinoCredentials;
 
 export const DbtProjectTypeLabels: Record<DbtProjectType, string> = {
     [DbtProjectType.DBT]: 'dbt local server',
@@ -759,102 +585,6 @@ export const DbtProjectTypeLabels: Record<DbtProjectType, string> = {
     [DbtProjectType.BITBUCKET]: 'BitBucket',
     [DbtProjectType.AZURE_DEVOPS]: 'Azure DevOps',
     [DbtProjectType.NONE]: 'CLI',
-};
-
-export interface DbtProjectConfigBase {
-    type: DbtProjectType;
-}
-
-export type DbtProjectEnvironmentVariable = {
-    key: string;
-    value: string;
-};
-
-export interface DbtProjectCompilerBase extends DbtProjectConfigBase {
-    target?: string;
-    environment?: DbtProjectEnvironmentVariable[];
-}
-
-export interface DbtNoneProjectConfig extends DbtProjectCompilerBase {
-    type: DbtProjectType.NONE;
-
-    hideRefreshButton?: boolean;
-}
-
-export interface DbtLocalProjectConfig extends DbtProjectCompilerBase {
-    type: DbtProjectType.DBT;
-    profiles_dir?: string;
-    project_dir?: string;
-}
-
-export interface DbtCloudIDEProjectConfig extends DbtProjectConfigBase {
-    type: DbtProjectType.DBT_CLOUD_IDE;
-    api_key: string;
-    account_id: string | number;
-    environment_id: string | number;
-    project_id: string | number;
-}
-
-export interface DbtGithubProjectConfig extends DbtProjectCompilerBase {
-    type: DbtProjectType.GITHUB;
-    personal_access_token: string;
-    repository: string;
-    branch: string;
-    project_sub_path: string;
-    host_domain?: string;
-}
-
-export interface DbtGitlabProjectConfig extends DbtProjectCompilerBase {
-    type: DbtProjectType.GITLAB;
-    personal_access_token: string;
-    repository: string;
-    branch: string;
-    project_sub_path: string;
-    host_domain?: string;
-}
-
-export interface DbtBitBucketProjectConfig extends DbtProjectCompilerBase {
-    type: DbtProjectType.BITBUCKET;
-    username: string;
-    personal_access_token: string;
-    repository: string;
-    branch: string;
-    project_sub_path: string;
-    host_domain?: string;
-}
-
-export interface DbtAzureDevOpsProjectConfig extends DbtProjectCompilerBase {
-    type: DbtProjectType.AZURE_DEVOPS;
-    personal_access_token: string;
-    organization: string;
-    project: string;
-    repository: string;
-    branch: string;
-    project_sub_path: string;
-}
-
-export type DbtProjectConfig =
-    | DbtLocalProjectConfig
-    | DbtCloudIDEProjectConfig
-    | DbtGithubProjectConfig
-    | DbtBitBucketProjectConfig
-    | DbtGitlabProjectConfig
-    | DbtAzureDevOpsProjectConfig
-    | DbtNoneProjectConfig;
-
-export type OrganizationProject = {
-    projectUuid: string;
-    name: string;
-    type: ProjectType;
-};
-
-export type Project = {
-    organizationUuid: string;
-    projectUuid: string;
-    name: string;
-    type: ProjectType;
-    dbtConnection: DbtProjectConfig;
-    warehouseConnection?: WarehouseCredentials;
 };
 
 export type CreateProject = Omit<
