+++ packages/frontend/src/components/ProjectConnection/DbtForms/GitlabForm.tsx
@@ -1,4 +1,3 @@
-import { DbtProjectType } from '@lightdash/common';
 import React, { FC } from 'react';
 import {
     hasNoWhiteSpaces,
