+++ packages/frontend/src/components/ProjectConnection/DbtSettingsForm.tsx
@@ -1,8 +1,4 @@
-import {
-    DbtProjectType,
-    DbtProjectTypeLabels,
-    WarehouseTypes,
-} from '@lightdash/common';
+import { DbtProjectTypeLabels } from '@lightdash/common';
 import { FC, useMemo, useState } from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
 import { useApp } from '../../providers/AppProvider';
