+++ packages/frontend/src/components/ProjectConnection/WarehouseForms/DatabricksForm.tsx
@@ -1,4 +1,3 @@
-import { WarehouseTypes } from '@lightdash/common';
 import React, { FC } from 'react';
 import { useToggle } from 'react-use';
 import { hasNoWhiteSpaces } from '../../../utils/fieldValidators';
