+++ packages/frontend/src/components/RefreshDbtButton/index.tsx
@@ -1,5 +1,5 @@
 import { Tooltip2 } from '@blueprintjs/popover2';
-import { DbtProjectType, ProjectType } from '@lightdash/common';
+import { ProjectType } from '@lightdash/common';
 import React, { ComponentProps, FC } from 'react';
 import { useParams } from 'react-router-dom';
 import { useProject } from '../../hooks/useProject';
